fix(bank-auto-login-lock): fix readability blockers in PR4A3 Redis lock wiring
- Reword defaults.ts singleton comment: 'Lazy singleton' was misleading
  because the singleton resolves at module evaluation, not lazily.
  The Redis socket stays lazy via lazyConnect: true.
- Extract shared FakeRedisClient test helper to eliminate duplication
  in defaults.test.ts. redis-store.test.ts retains its own copy for
  now to stay under the 400-line PR budget; a follow-up can unify.
