+++ src/modules/bank-auto-login-lock/defaults.test.ts
@@ -1,73 +1,6 @@
 import { describe, it, expect, beforeEach, afterEach, vi } from "vitest";
-import type { RedisEvalClient } from "./redis-store";
 import { createAutoLoginLockFromEnv } from "./defaults";
-
-// ---------------------------------------------------------------------------
-// Fake Redis client — same contract as redis-store.test.ts
-// ---------------------------------------------------------------------------
-
-class FakeRedisClient implements RedisEvalClient {
-  private store = new Map<string, string>();
-  private expiry = new Map<string, number>();
-  private fenceCounters = new Map<string, number>();
-  public now = Date.now();
-
-  private isExpired(key: string): boolean {
-    const exp = this.expiry.get(key);
-    if (exp == null) return false;
-    if (this.now >= exp) {
-      this.store.delete(key);
-      this.expiry.delete(key);
-      return true;
-    }
-    return false;
-  }
-
-  async eval(script: string, numKeys: number, ...args: (string | number)[]): Promise<unknown> {
-    const keys = args.slice(0, numKeys) as string[];
-    const scriptArgs = args.slice(numKeys);
-
-    const isAcquire = script.includes("INCR") && script.includes("fenceKey");
-    const isRelease = script.includes("expectedToken") && script.includes("DEL") && !script.includes("INCR");
-    const isRenew = script.includes("newTtlMs") && script.includes("SET") && !script.includes("INCR");
-
-    if (isAcquire) {
-      const [lockKey, fenceKey, leaseToken, ttlMs] = [keys[0], keys[1], scriptArgs[0] as string, Number(scriptArgs[1])];
-      if (!this.isExpired(lockKey) && this.store.has(lockKey)) return null;
-      const next = (this.fenceCounters.get(fenceKey) ?? 0) + 1;
-      this.fenceCounters.set(fenceKey, next);
-      this.store.set(lockKey, JSON.stringify({ leaseToken, fencingToken: next }));
-      this.expiry.set(lockKey, this.now + ttlMs);
-      return next;
-    }
-
-    if (isRelease) {
-      const expected = scriptArgs[0] as string;
-      if (this.isExpired(keys[0])) return 0;
-      const raw = this.store.get(keys[0]);
-      if (!raw) return 0;
-      const d = JSON.parse(raw) as { leaseToken: string };
-      if (d.leaseToken !== expected) return 0;
-      this.store.delete(keys[0]);
-      this.expiry.delete(keys[0]);
-      return 1;
-    }
-
-    if (isRenew) {
-      const [expected, ttlMs] = [scriptArgs[0] as string, Number(scriptArgs[1])];
-      if (this.isExpired(keys[0])) return 0;
-      const raw = this.store.get(keys[0]);
-      if (!raw) return 0;
-      const d = JSON.parse(raw) as { leaseToken: string; fencingToken: number; expiresAt: number };
-      if (d.leaseToken !== expected) return 0;
-      this.expiry.set(keys[0], this.now + ttlMs);
-      this.store.set(keys[0], JSON.stringify({ ...d, expiresAt: this.now + ttlMs }));
-      return 1;
-    }
-
-    throw new Error("Unexpected Lua script");
-  }
-}
+import { FakeRedisClient } from "./fake-redis-client";
 
 // ---------------------------------------------------------------------------
 // Test suite
